Launcher2 multi-profile support
Use LauncherApps API and badging APIs instead of PackageManager.

Adds support to show apps from current user and any managed profiles.

Background: Managed profiles are user sandboxes that are visible from
the primary user and can be launched as if they are a part of this user.
A launcher should now be capable of listing apps from this user as well
as related profiles of this user.

Launching of activities is now via the LauncherApps interface, to allow
for cross-profile app launching. Only activities with category LAUNCHER
can be added as a shortcut on the workspace for a managed profile.

Widgets are only supported for the current profile. Widgets from the
managed profile are not available.

TODO: Handle users appearing and disappearing.

Change-Id: I732a476d68236f7de7eaa5a2105c868621c8a57e