Always load and bind all apps last after launcher reload.
Previously, all apps view could be bound first, but when
a managed profile was added this led to the call to refresh
the view being nobbled by the workspace refresh and certain
applications not being displayed.

Always loading the all apps view last is a safe workaround and
mirrors the change made to Launcher3. See bug for more details.

Bug: 17336902
Change-Id: I7c4ab1e21e45d397f98583709ebb09e069fa2be9